requirements.txt
beautifulsoup4
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
langchain-community
playwright==1.49.0
python-dotenv
pypdf
Flask
setuptools
langtrace-python-sdk